pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "traffic-env"
version = "1.0.0"
description = "Traffic Queue RL Environment"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "openai",
    "httpx",
    "python-dotenv",
    "openenv-core>=0.1.0"
]

[project.scripts]
server = "server.app:main"